+++ src/lib/icons.ts
@@ -38,6 +38,7 @@ export { default as IconMeter } from "./icons/IconMeter.svelte";
 export { default as IconNeurons } from "./icons/IconNeurons.svelte";
 export { default as IconNeuronsPage } from "./icons/IconNeuronsPage.svelte";
 export { default as IconNorthEast } from "./icons/IconNorthEast.svelte";
+export { default as IconOpenInNew } from "./icons/IconOpenInNew.svelte";
 export { default as IconPace } from "./icons/IconPace.svelte";
 export { default as IconPassword } from "./icons/IconPassword.svelte";
 export { default as IconQRCode } from "./icons/IconQRCode.svelte";
@@ -62,4 +63,3 @@ export { default as IconVote } from "./icons/IconVote.svelte";
 export { default as IconWallet } from "./icons/IconWallet.svelte";
 export { default as IconWarning } from "./icons/IconWarning.svelte";
 export { default as IconWest } from "./icons/IconWest.svelte";
-export { default as IconOpenInNew } from "./icons/OpenInNew.svelte";
