+++ src/lib/icons/IconOpenInNew.svelte
